trigger confetti inside TransactionStatusProvider
