+++ src/components/[guild]/Requirements/components/GuildCheckout/components/TransactionStatusContext.tsx
@@ -1,10 +1,12 @@
 import { useDisclosure } from "@chakra-ui/react"
+import useJsConfetti from "components/create-guild/hooks/useJsConfetti"
 import {
   createContext,
   Dispatch,
   PropsWithChildren,
   SetStateAction,
   useContext,
+  useEffect,
   useState,
 } from "react"
 
@@ -32,6 +34,13 @@ const TransactionStatusProvider = ({
   const [txHash, setTxHash] = useState("")
   const [txError, setTxError] = useState(false)
   const [txSuccess, setTxSuccess] = useState(false)
+
+  const triggerConfetti = useJsConfetti()
+
+  useEffect(() => {
+    if (!txSuccess) return
+    triggerConfetti()
+  }, [txSuccess])
 
   return (
     <TransactionStatusContext.Provider
